fixed the seizures
